+++ docs/src/components/PlaygroundSettingsInputs.module.css
@@ -1,26 +1,6 @@
-.content {
-    zoom: 1.28;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    position: absolute;
-    background-color: #242427;
+.container {
     display: flex;
     flex-wrap: wrap;
-    padding: 20px;
-    border-radius: 4px;
-    border: 1px solid #444950;
-    width: 548px;
-    margin: 0 auto;
-}
-
-.overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: rgba(0, 0, 0, 0.75);
 }
 
 .inputSelector {
@@ -49,10 +29,9 @@
     flex-basis: 100%;
     display: flex;
     margin-top: 6px;
-    justify-content: flex-end;
+    justify-content: center;
 }
 
 .controlButton {
-    margin: 0 20px 0 0px;
-    width: 96px;
+    width: 128px;
 }
